feat(infra): finish singleVm
