+++ cdc/package.json
@@ -3,6 +3,9 @@
   "version": "0.0.2",
   "description": "PostgreSQL logical-replication worker for activity and sync events",
   "type": "module",
+  "exports": {
+    ".": "./src/index.ts"
+  },
   "engines": {
     "node": "24.x"
   },
