+++ package.json
@@ -30,6 +30,7 @@
     "sdk": "pnpm --filter backend generate:openapi && pnpm --filter sdk generate:sdk",
     "clean": "rimraf backend/dist frontend/dist frontend/stats",
     "dev": "pnpm clean && cross-env pnpm -r --parallel --stream dev",
+    "dev:single": "pnpm clean && cross-env SINGLE_VM=true pnpm -r --parallel --stream --filter backend --filter frontend dev",
     "dev:migration": "pnpm clean && cross-env DATABASE_URL=postgres://postgres:postgres@0.0.0.0:5555/postgres DATABASE_ADMIN_URL=postgres://postgres:postgres@0.0.0.0:5555/postgres pnpm -r --parallel --stream dev",
     "tunnel": "pnpm clean && cross-env pnpm -r --parallel --stream tunnel",
     "dev:proxy": "pnpm clean && cross-env pnpm -r --parallel --stream --filter backend --filter cdc --filter frontend --filter proxy dev",
@@ -40,7 +41,7 @@
     "lint": "pnpm biome check .",
     "lint:fix": "pnpm biome check --write --unsafe",
     "ts": "pnpm -r ts",
-    "tsperf": "pnpm --filter backend tsperf && pnpm --filter frontend tsperf && pnpm --filter cdc tsperf && pnpm --filter yjs tsperf",
+    "tsperf": "pnpm --filter backend tsperf && pnpm --filter frontend tsperf && pnpm --filter cdc tsperf && pnpm --filter yjs-worker tsperf",
     "build": "pnpm -r --parallel clean && pnpm -r --parallel build",
     "build:dev": "pnpm -r --parallel build:dev",
     "build:staging": "pnpm -r --parallel build:staging",
